Moving dolt_verify_constraints script tests over to their own file
